new "components" folder structure
