--- src/components/NoteItem.module.scss
@@ -1,36 +1,0 @@
-.heading {
-  font-weight: bold;
-}
-
-.text {
-  padding-right: 1rem;
-  margin-bottom: 1rem;
-  font-size: 0.825rem;
-}
-
-.textHidden {
-  composes: text;
-  width: 100%;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.details {
-  font-size: 0.825rem;
-  margin-bottom: 1rem;
-}
-
-.tags {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 0.25rem;
-}
-
-.actions {
-  display: flex;
-  justify-content: flex-end;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
