+++ src/components/notes/NoteItem/NoteItemActions.module.scss
@@ -1,0 +1,6 @@
+.actions {
+  display: flex;
+  justify-content: flex-end;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
