+++ src/components/notes/NoteItem/NoteItemDetails.module.scss
@@ -1,0 +1,12 @@
+.details {
+  font-size: 0.825rem;
+  margin-bottom: 1rem;
+}
+
+.tags {
+  margin-top: 0.25rem;
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: 0.25rem;
+}
